+++ src/main/java/com/uber/cadence/internal/dispatcher/SyncDecisionContext.java
@@ -25,9 +25,9 @@ import com.uber.cadence.converter.DataConverter;
 import com.uber.cadence.internal.ActivityException;
 import com.uber.cadence.internal.ChildWorkflowTaskFailedException;
 import com.uber.cadence.internal.WorkflowRetryerInternal;
-import com.uber.cadence.internal.generic.ExecuteActivityParameters;
-import com.uber.cadence.internal.worker.ActivityTaskTimeoutException;
-import com.uber.cadence.internal.worker.DecisionContext;
+import com.uber.cadence.internal.replay.ExecuteActivityParameters;
+import com.uber.cadence.internal.replay.ActivityTaskTimeoutException;
+import com.uber.cadence.internal.replay.DecisionContext;
 import com.uber.cadence.internal.worker.POJOQueryImplementationFactory;
 import com.uber.cadence.workflow.ActivityFailureException;
 import com.uber.cadence.workflow.ActivityTimeoutException;
@@ -35,10 +35,10 @@ import com.uber.cadence.workflow.CancellationScope;
 import com.uber.cadence.workflow.ChildWorkflowFailureException;
 import com.uber.cadence.workflow.ChildWorkflowOptions;
 import com.uber.cadence.workflow.CompletablePromise;
-import com.uber.cadence.workflow.ContinueAsNewWorkflowExecutionParameters;
+import com.uber.cadence.internal.replay.ContinueAsNewWorkflowExecutionParameters;
 import com.uber.cadence.workflow.Functions;
 import com.uber.cadence.workflow.Promise;
-import com.uber.cadence.workflow.StartChildWorkflowExecutionParameters;
+import com.uber.cadence.internal.replay.StartChildWorkflowExecutionParameters;
 import com.uber.cadence.workflow.Workflow;
 
 import java.util.HashMap;
