+++ src/main/java/com/uber/cadence/internal/dispatcher/SyncWorkflowWorker.java
@@ -19,7 +19,7 @@ package com.uber.cadence.internal.dispatcher;
 import com.uber.cadence.WorkflowExecution;
 import com.uber.cadence.WorkflowService;
 import com.uber.cadence.converter.DataConverter;
-import com.uber.cadence.internal.worker.ReplayDecisionTaskHandler;
+import com.uber.cadence.internal.replay.ReplayDecisionTaskHandler;
 import com.uber.cadence.internal.worker.AsyncWorkflowFactory;
 import com.uber.cadence.internal.worker.DecisionTaskPoller;
 import com.uber.cadence.internal.worker.GenericWorker;
