+++ src/main/java/com/uber/cadence/internal/replay/ChildWorkflowDecisionStateMachine.java
@@ -14,15 +14,18 @@
  *  express or implied. See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package com.uber.cadence.internal.worker;
+package com.uber.cadence.internal.replay;
 
 import com.uber.cadence.Decision;
 import com.uber.cadence.DecisionType;
 import com.uber.cadence.HistoryEvent;
 import com.uber.cadence.RequestCancelExternalWorkflowExecutionDecisionAttributes;
 import com.uber.cadence.StartChildWorkflowExecutionDecisionAttributes;
+import com.uber.cadence.internal.replay.DecisionId;
+import com.uber.cadence.internal.replay.DecisionState;
+import com.uber.cadence.internal.replay.DecisionStateMachineBase;
 
-class ChildWorkflowDecisionStateMachine extends DecisionStateMachineBase {
+final class ChildWorkflowDecisionStateMachine extends DecisionStateMachineBase {
 
     private StartChildWorkflowExecutionDecisionAttributes startAttributes;
