+++ src/main/java/com/uber/cadence/internal/replay/DecisionContext.java
@@ -14,13 +14,10 @@
  *  express or implied. See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package com.uber.cadence.internal.worker;
+package com.uber.cadence.internal.replay;
 
 import com.uber.cadence.WorkflowExecution;
 import com.uber.cadence.WorkflowType;
-import com.uber.cadence.internal.generic.ExecuteActivityParameters;
-import com.uber.cadence.workflow.ContinueAsNewWorkflowExecutionParameters;
-import com.uber.cadence.workflow.StartChildWorkflowExecutionParameters;
 
 import java.time.Duration;
 import java.util.function.BiConsumer;
