+++ src/main/java/com/uber/cadence/internal/replay/DecisionContextImpl.java
@@ -14,7 +14,7 @@
  *  express or implied. See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package com.uber.cadence.internal.worker;
+package com.uber.cadence.internal.replay;
 
 import com.uber.cadence.ActivityTaskStartedEventAttributes;
 import com.uber.cadence.HistoryEvent;
@@ -23,15 +23,12 @@ import com.uber.cadence.TimerFiredEventAttributes;
 import com.uber.cadence.WorkflowExecution;
 import com.uber.cadence.WorkflowExecutionStartedEventAttributes;
 import com.uber.cadence.WorkflowType;
-import com.uber.cadence.internal.generic.ExecuteActivityParameters;
-import com.uber.cadence.workflow.ContinueAsNewWorkflowExecutionParameters;
-import com.uber.cadence.workflow.StartChildWorkflowExecutionParameters;
 
 import java.time.Duration;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 
-class DecisionContextImpl implements DecisionContext, HistoryEventHandler {
+final class DecisionContextImpl implements DecisionContext, HistoryEventHandler {
 
     private final ActivityDecisionContext activityClient;
