+++ src/main/java/com/uber/cadence/internal/replay/HistoryHelper.java
@@ -14,13 +14,14 @@
  *  express or implied. See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package com.uber.cadence.internal.worker;
+package com.uber.cadence.internal.replay;
 
 import com.uber.cadence.WorkflowExecutionStartedEventAttributes;
 import com.uber.cadence.internal.common.WorkflowExecutionUtils;
 import com.uber.cadence.EventType;
 import com.uber.cadence.HistoryEvent;
 import com.uber.cadence.PollForDecisionTaskResponse;
+import com.uber.cadence.internal.worker.DecisionTaskWithHistoryIterator;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
