+++ src/main/java/com/uber/cadence/internal/replay/OpenRequestInfo.java
@@ -14,7 +14,7 @@
  *  express or implied. See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package com.uber.cadence.internal.worker;
+package com.uber.cadence.internal.replay;
 
 import java.util.function.BiConsumer;
