+++ src/main/java/com/uber/cadence/internal/replay/SignalExternalWorkflowParameters.java
@@ -14,19 +14,19 @@
  *  express or implied. See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package com.uber.cadence.internal.generic;
+package com.uber.cadence.internal.replay;
 
-public class QueryWorkflowParameters implements Cloneable {
+public final class SignalExternalWorkflowParameters implements Cloneable {
 
     private byte[] input;
-
+    
     private String runId;
-
-    private String queryType;
-
+    
+    private String signalName;
+    
     private String workflowId;
-
-    public QueryWorkflowParameters() {
+    
+    public SignalExternalWorkflowParameters() {
     }
     
     public byte[] getInput() {
@@ -37,7 +37,7 @@ public class QueryWorkflowParameters implements Cloneable {
         this.input = input;
     }
     
-    public QueryWorkflowParameters withInput(byte[] input) {
+    public SignalExternalWorkflowParameters withInput(byte[] input) {
         this.input = input;
         return this;
     }
@@ -50,21 +50,21 @@ public class QueryWorkflowParameters implements Cloneable {
         this.runId = runId;
     }
     
-    public QueryWorkflowParameters withRunId(String runId) {
+    public SignalExternalWorkflowParameters withRunId(String runId) {
         this.runId = runId;
         return this;
     }
 
-    public String getQueryType() {
-        return queryType;
+    public String getSignalName() {
+        return signalName;
     }
     
-    public void setQueryType(String queryType) {
-        this.queryType = queryType;
+    public void setSignalName(String signalName) {
+        this.signalName = signalName;
     }
     
-    public QueryWorkflowParameters withQueryType(String queryType) {
-        this.queryType = queryType;
+    public SignalExternalWorkflowParameters withSignalName(String signalName) {
+        this.signalName = signalName;
         return this;
     }
     
@@ -76,7 +76,7 @@ public class QueryWorkflowParameters implements Cloneable {
         this.workflowId = workflowId;
     }
     
-    public QueryWorkflowParameters withWorkflowId(String workflowId) {
+    public SignalExternalWorkflowParameters withWorkflowId(String workflowId) {
         this.workflowId = workflowId;
         return this;
     }
@@ -85,7 +85,7 @@ public class QueryWorkflowParameters implements Cloneable {
     public String toString() {
         StringBuilder sb = new StringBuilder();
         sb.append("{");
-        sb.append("QueryName: " + queryType + ", ");
+        sb.append("SignalName: " + signalName + ", ");
         sb.append("Input: " + input + ", ");
         sb.append("WorkflowId: " + workflowId + ", ");
         sb.append("RunId: " + runId + ", ");
@@ -93,11 +93,11 @@ public class QueryWorkflowParameters implements Cloneable {
         return sb.toString();
     }
 
-    public QueryWorkflowParameters clone() {
-        QueryWorkflowParameters result = new QueryWorkflowParameters();
+    public SignalExternalWorkflowParameters clone() {
+        SignalExternalWorkflowParameters result = new SignalExternalWorkflowParameters();
         result.setInput(input);
         result.setRunId(runId);
-        result.setQueryType(queryType);
+        result.setSignalName(signalName);
         result.setWorkflowId(workflowId);
         return result;
     }
